// src/app/Models/LocalStorageObject.ts
export interface LocalStorageObject {
    VerbEndings: string[];
    Tenses: string[];
    AuxVerbs: string[];
    PresentVerbs: string[];
    PasseComposeVerbs: string[];
    FutureVerbs: string[];
    ImparfaitVerbs: string[];
    Pronouns: string[];
    AlwaysShowHint: boolean;
    EnglishTrans: boolean;
}     